refactor: api instance
